Moved post item to seperate react comp
